requirements.txt
python-dotenv# API Telegrampython-telegram-bot==21.5# HTTPrequestshttpx>=0.27.0cfficryptography# Manip donnéespydanticdataclasses-jsontyping-extensions# LangChainlangchain-corelangchain-openailangchain-community# RAG (Base de données)chromadb==0.5.3duckdbuvicorn# RAG (Modèles d'Embedding) - C'ÉTAIT LE MANQUANT CRITIQUE# UtilitairesalembicpsutilPyYAML